+++ src/core/base.jl
@@ -1,14 +1,5 @@
 # stuff that is universal to all power models
 
-export
-    GenericPowerModel,
-    optimize!,
-    run_generic_model, build_generic_model, solve_generic_model,
-    ismultinetwork, nw_ids, nws,
-    ismulticonductor, conductor_ids,
-    ids, ref, var, con, ext
-
-
 "root of the power formulation type hierarchy"
 abstract type AbstractPowerFormulation end
 
@@ -67,7 +58,7 @@ function GenericPowerModel(data::Dict{String,<:Any}, T::DataType; ext = Dict{Sym
     # TODO is may be a good place to check component connectivity validity
     # i.e. https://github.com/lanl-ansi/PowerModels.jl/issues/131
 
-    ref = build_generic_ref(data) # refrence data
+    ref = _build_generic_ref(data) # refrence data
 
     var = Dict{Symbol,Any}(:nw => Dict{Int,Any}())
     con = Dict{Symbol,Any}(:nw => Dict{Int,Any}())
@@ -163,17 +154,17 @@ con(pm::GenericPowerModel, key::Symbol, idx; nw::Int=pm.cnw, cnd::Int=pm.ccnd) =
 
 
 function JuMP.optimize!(pm::GenericPowerModel, optimizer::JuMP.OptimizerFactory)
-    if pm.model.moi_backend.state == MOIU.NO_OPTIMIZER
+    if pm.model.moi_backend.state == _MOI.Utilities.NO_OPTIMIZER
         _, solve_time, solve_bytes_alloc, sec_in_gc = @timed JuMP.optimize!(pm.model, optimizer)
     else
-        Memento.warn(LOGGER, "Model already contains optimizer factory, cannot use optimizer specified in `solve_generic_model`")
+        Memento.warn(_LOGGER, "Model already contains optimizer factory, cannot use optimizer specified in `solve_generic_model`")
         _, solve_time, solve_bytes_alloc, sec_in_gc = @timed JuMP.optimize!(pm.model)
     end
 
     try
-        solve_time = MOI.get(pm.model, MOI.SolveTime())
+        solve_time = _MOI.get(pm.model, _MOI.SolveTime())
     catch
-        Memento.warn(LOGGER, "the given optimizer does not provide the SolveTime() attribute, falling back on @timed.  This is not a rigorous timing value.");
+        Memento.warn(_LOGGER, "the given optimizer does not provide the SolveTime() attribute, falling back on @timed.  This is not a rigorous timing value.");
     end
 
     return solve_time
@@ -189,11 +180,11 @@ end
 function run_generic_model(data::Dict{String,<:Any}, model_constructor, optimizer, post_method; ref_extensions=[], solution_builder=get_solution, kwargs...)
     #start_time = time()
     pm = build_generic_model(data, model_constructor, post_method; ref_extensions=ref_extensions, kwargs...)
-    #Memento.info(LOGGER, "pm model build time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "pm model build time: $(time() - start_time)")
 
     #start_time = time()
     solution = solve_generic_model(pm, optimizer; solution_builder = solution_builder)
-    #Memento.info(LOGGER, "pm model solve and solution time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "pm model solve and solution time: $(time() - start_time)")
 
     return solution
 end
@@ -210,26 +201,26 @@ function build_generic_model(data::Dict{String,<:Any}, model_constructor, post_m
 
     #start_time = time()
     pm = model_constructor(data; kwargs...)
-    #Memento.info(LOGGER, "pm model_constructor time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "pm model_constructor time: $(time() - start_time)")
 
     if !multinetwork && ismultinetwork(pm)
-        Memento.error(LOGGER, "attempted to build a single-network model with multi-network data")
+        Memento.error(_LOGGER, "attempted to build a single-network model with multi-network data")
     end
 
     if !multiconductor && ismulticonductor(pm)
-        Memento.error(LOGGER, "attempted to build a single-conductor model with multi-conductor data")
+        Memento.error(_LOGGER, "attempted to build a single-conductor model with multi-conductor data")
     end
 
     #start_time = time()
-    core_ref!(pm)
+    ref_core!(pm)
     for ref_ext in ref_extensions
         ref_ext(pm)
     end
-    #Memento.info(LOGGER, "pm build ref time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "pm build ref time: $(time() - start_time)")
 
     #start_time = time()
     post_method(pm)
-    #Memento.info(LOGGER, "pm post_method time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "pm post_method time: $(time() - start_time)")
 
     return pm
 end
@@ -240,11 +231,11 @@ function solve_generic_model(pm::GenericPowerModel, optimizer::JuMP.OptimizerFac
 
     #start_time = time()
     solve_time = JuMP.optimize!(pm, optimizer)
-    #Memento.info(LOGGER, "JuMP model optimize time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "JuMP model optimize time: $(time() - start_time)")
 
     #start_time = time()
     solution = build_solution(pm, solve_time; solution_builder = solution_builder)
-    #Memento.info(LOGGER, "PowerModels solution build time: $(time() - start_time)")
+    #Memento.info(_LOGGER, "PowerModels solution build time: $(time() - start_time)")
 
     return solution
 end
@@ -252,8 +243,8 @@ end
 
 "used for building ref without the need to build a GenericPowerModel"
 function build_ref(data::Dict{String,<:Any}; ref_extensions=[])
-    ref = build_generic_ref(data)
-    core_ref!(ref[:nw])
+    ref = _build_generic_ref(data)
+    _ref_core!(ref[:nw])
     for ref_ext in ref_extensions
         ref_ext(pm)
     end
@@ -261,7 +252,7 @@ function build_ref(data::Dict{String,<:Any}; ref_extensions=[])
 end
 
 
-function build_generic_ref(data::Dict{String,<:Any})
+function _build_generic_ref(data::Dict{String,<:Any})
     refs = Dict{Symbol,Any}()
 
     nws = refs[:nw] = Dict{Int,Any}()
@@ -325,11 +316,11 @@ If `:ne_branch` exists, then the following keys are also available with similar 
 
 * `:ne_branch`, `:ne_arcs_from`, `:ne_arcs_to`, `:ne_arcs`, `:ne_bus_arcs`, `:ne_buspairs`.
 """
-function core_ref!(pm::GenericPowerModel)
-    core_ref!(pm.ref[:nw])
+function ref_core!(pm::GenericPowerModel)
+    _ref_core!(pm.ref[:nw])
 end
 
-function core_ref!(nw_refs::Dict)
+function _ref_core!(nw_refs::Dict)
     for (nw, ref) in nw_refs
 
         ### filter out inactive components ###
@@ -401,7 +392,7 @@ function core_ref!(nw_refs::Dict)
         ref[:ref_buses] = ref_buses
 
         if length(ref_buses) > 1
-            Memento.warn(LOGGER, "multiple reference buses found, $(keys(ref_buses)), this can cause infeasibility if they are in the same connected component")
+            Memento.warn(_LOGGER, "multiple reference buses found, $(keys(ref_buses)), this can cause infeasibility if they are in the same connected component")
         end
 
 
@@ -414,7 +405,7 @@ end
 
 
 "checks of any of the given keys are missing from the given dict"
-function check_missing_keys(dict, keys, type)
+function _check_missing_keys(dict, keys, type)
     missing = []
     for key in keys
         if !haskey(dict, key)
@@ -422,7 +413,7 @@ function check_missing_keys(dict, keys, type)
         end
     end
     if length(missing) > 0
-        error(LOGGER, "the formulation $(type) requires the following varible(s) $(keys) but the $(missing) variable(s) were not found in the model")
+        error(_LOGGER, "the formulation $(type) requires the following varible(s) $(keys) but the $(missing) variable(s) were not found in the model")
     end
 end