+++ src/form/shared.jl
@@ -170,7 +170,7 @@ end
 
 
 ""
-function add_bus_voltage_setpoint(sol, pm::GenericPowerModel{T}) where T <: AbstractWForms
+function add_setpoint_bus_voltage(sol, pm::GenericPowerModel{T}) where T <: AbstractWForms
     add_setpoint(sol, pm, "bus", "vm", :w; scale = (x,item,cnd) -> sqrt(x))
     # What should the default value be?
     add_setpoint_fixed(sol, pm, "bus", "va")