+++ src/prob/opf.jl
@@ -1,5 +1,3 @@
-export run_opf, run_ac_opf, run_dc_opf
-
 ""
 function run_ac_opf(file, optimizer; kwargs...)
     return run_opf(file, ACPPowerModel, optimizer; kwargs...)
@@ -48,3 +46,47 @@ function post_opf(pm::GenericPowerModel)
         constraint_dcline(pm, i)
     end
 end
+
+
+
+"a toy example of how to model with multi-networks"
+function run_mn_opf(file, model_constructor, optimizer; kwargs...)
+    return run_generic_model(file, model_constructor, optimizer, post_mn_opf; multinetwork=true, kwargs...)
+end
+
+""
+function post_mn_opf(pm::GenericPowerModel)
+    for (n, network) in nws(pm)
+        variable_voltage(pm, nw=n)
+        variable_generation(pm, nw=n)
+        variable_branch_flow(pm, nw=n)
+        variable_dcline_flow(pm, nw=n)
+
+        constraint_model_voltage(pm, nw=n)
+
+        for i in ids(pm, :ref_buses, nw=n)
+            constraint_theta_ref(pm, i, nw=n)
+        end
+
+        for i in ids(pm, :bus, nw=n)
+            constraint_kcl_shunt(pm, i, nw=n)
+        end
+
+        for i in ids(pm, :branch, nw=n)
+            constraint_ohms_yt_from(pm, i, nw=n)
+            constraint_ohms_yt_to(pm, i, nw=n)
+
+            constraint_voltage_angle_difference(pm, i, nw=n)
+
+            constraint_thermal_limit_from(pm, i, nw=n)
+            constraint_thermal_limit_to(pm, i, nw=n)
+        end
+
+        for i in ids(pm, :dcline, nw=n)
+            constraint_dcline(pm, i, nw=n)
+        end
+    end
+
+    objective_min_fuel_and_flow_cost(pm)
+end
+