+++ test/multi-nw-cnd.jl
@@ -114,9 +114,9 @@
 
             @test length(mn_mc_data["nw"]) == 3
 
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 157967.0; atol = 1e0)
 
             @test length(result["solution"]["nw"]) == 3
@@ -159,9 +159,9 @@
 
             @test length(mn_mc_data["nw"]) == 2
 
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 2.67529e5; atol = 1e0)
 
             @test length(result["solution"]["nw"]) == 2
@@ -176,9 +176,9 @@
 
             @test length(mn_mc_data["nw"]) == 2
 
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 120623.0; atol = 1e1)
 
             @test length(result["solution"]["nw"]) == 2
@@ -198,23 +198,23 @@
         mn_mc_data = build_mn_mc_data("../test/data/matpower/case5_dc.m", "../test/data/matpower/case14.m", conductors_1=4, conductors_2=0)
 
         @testset "ac 5/14-bus case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 80706.2; atol = 1e-1)
         end
 
         @testset "dc 5/14-bus case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, DCPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, DCPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 80006.2; atol = 1e-1)
         end
 
         @testset "soc 5/14-bus case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, SOCWRPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, SOCWRPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 69827.3; atol = 1e-1)
         end
 
@@ -225,33 +225,33 @@
         mn_mc_data = build_mn_mc_data("../test/data/matpower/case5_strg.m")
 
         @testset "ac 5-bus storage case" begin
-            result = PowerModels.run_mn_mc_strg_opf(mn_mc_data, PowerModels.ACPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_strg_opf(mn_mc_data, PowerModels.ACPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 1.35456e5; atol = 1e2)
         end
 
         @testset "dc 5-bus storage case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, PowerModels.DCPPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, PowerModels.DCPPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 1.58593e5; atol = 1e2)
         end
 
         #=
         # base case not yet implemented
         @testset "soc 5-bus storage case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, SOCWRPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, SOCWRPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 69827.3; atol = 1e-1)
         end
         =#
 
         @testset "nfa 5 bus storage case" begin
-            result = PowerModels.run_mn_mc_opf(mn_mc_data, PowerModels.NFAPowerModel, ipopt_solver)
+            result = PowerModels._run_mn_mc_opf(mn_mc_data, PowerModels.NFAPowerModel, ipopt_solver)
 
-            @test result["termination_status"] == MOI.LOCALLY_SOLVED
+            @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 1.33302e5; atol = 1e2)
         end
     end
@@ -260,9 +260,9 @@
     @testset "test solution feedback" begin
         mn_mc_data = build_mn_mc_data("../test/data/matpower/case5_dc.m", "../test/data/matpower/case5_asym.m", conductors_1=4, conductors_2=0)
 
-        result = PowerModels.run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
+        result = PowerModels._run_mn_mc_opf(mn_mc_data, ACPPowerModel, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 90176.6; atol = 1e0)
 
         PowerModels.update_data!(mn_mc_data, result["solution"])