+++ test/opf-obj.jl
@@ -9,21 +9,21 @@
     @testset "nlp solver" begin
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 5420.3; atol = 1e0)
     end
 
     @testset "conic solver" begin
         result = run_opf(data, SOCWRConicPowerModel, scs_solver)
 
-        @test result["termination_status"] == MOI.OPTIMAL
+        @test result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 3095.88; atol = 1e0)
     end
 
     @testset "lp solver" begin
         result = run_dc_opf(data, cbc_solver)
 
-        @test result["termination_status"] == MOI.OPTIMAL
+        @test result["termination_status"] == OPTIMAL
         # @test isapprox(result["objective"], 4679.05; atol = 1e0)  # Problem upstream with JuMP.SecondOrderCone or JuMP.RotatedSecondOrderCone?
     end
 end
@@ -37,14 +37,14 @@ end
     @testset "nlp solver" begin
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 5420.46; atol = 1e0)
     end
 
     @testset "conic solver" begin
         result = run_opf(data, SOCWRConicPowerModel, scs_solver)
 
-        @test result["termination_status"] == MOI.OPTIMAL
+        @test result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 3096.04; atol = 1e0)
     end
 end
@@ -58,14 +58,14 @@ end
     @testset "opf objective" begin
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 5458.52; atol = 1e0)
     end
 
     @testset "opb objective" begin
         result = run_opb(data, SOCWRPowerModel, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 2962.22; atol = 1e0)
     end
 
@@ -79,7 +79,7 @@ end
         data["dcline"]["1"]["cost"] = [1.0, 1.0, 4000.0, 1.0]
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 18197.2; atol = 1e0)
     end
 
@@ -87,7 +87,7 @@ end
         data["dcline"]["1"]["cost"] = [1.0, 4000.0, 1.0]
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 18157.2; atol = 1e0)
     end
 
@@ -95,7 +95,7 @@ end
         data["dcline"]["1"]["cost"] = [4000.0, 1.0]
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 18157.2; atol = 1e0)
     end
 
@@ -103,7 +103,7 @@ end
         data["dcline"]["1"]["cost"] = [1.0]
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 17757.2; atol = 1e0)
     end
 
@@ -111,7 +111,7 @@ end
         data["dcline"]["1"]["cost"] = []
         result = run_ac_opf(data, ipopt_solver)
 
-        @test result["termination_status"] == MOI.LOCALLY_SOLVED
+        @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 17756.2; atol = 1e0)
     end
 end